dorabky k obhajobe - rozsirenie properties, zobrazovanie na mape

+++ paz1cProjekt/src/main/java/persistent/AddressDao.java
@@ -11,5 +11,5 @@ public interface AddressDao {
 
 	Address save(Address address);
 	
-	Address getById(int id);
+	Address getById(Long id);
 }
